Rename tests and folder
