+++ cypress/e2e/smokeTest/ProductSearch.cy.ts
