+++ cypress/e2e/smokeTest/VerifyHomepageLoads.cy.ts
@@ -1,4 +1,4 @@
-import homePageSelectors from '../../fixtures/verify_Homepage_Loads.json'
+import homePageSelectors from '../../fixtures/verifyHomepageLoadsSelectors.json'
 describe('Verify Homepage Loads Successfully', () => {
 
     // Before the test begins, navigate to the homepage
