+++ cypress/fixtures/addProductToTheCartSelectors.json
