+++ cypress/fixtures/verifyHomepageLoadsSelectors.json
